Refactor core module into granular, focused submodules
